# Improve performance of retrieving blob data