--- a/lib/rjgit.rb
+++ b/lib/rjgit.rb
@@ -56,18 +56,11 @@ module RJGit
 
     # http://dev.eclipse.org/mhonarc/lists/jgit-dev/msg00558.html
     def self.cat_file(repository, blob)
-      path = blob.path if blob.respond_to?(:path)
+      mode = blob.mode if blob.respond_to?(:mode)
       jrepo = RJGit.repository_type(repository)
       jblob = RJGit.blob_type(blob)
       # Try to resolve symlinks; return nil otherwise
-      mode = if path
-        last_commit_hash = jrepo.resolve(Constants::HEAD)
-        return nil unless last_commit_hash
-        jtree = RevWalk.new(jrepo).parse_commit(last_commit_hash).get_tree
-        RJGit.get_file_mode_with_path(jrepo, path, jtree)
-      else
-        RJGit.get_file_mode(jrepo, jblob)
-      end
+      mode ||= RJGit.get_file_mode(jrepo, jblob)
       if mode == SYMLINK_TYPE
         symlink_source = jrepo.open(jblob.id).get_bytes.to_a.pack('c*').force_encoding('UTF-8')
         blob = Blob.find_blob(jrepo, symlink_source)
@@ -221,7 +214,8 @@ module RJGit
       # To avoid missing full-line matches during the optimization, we first convert multiline anchors to single-line anchors.
       query = Regexp.new(query.source.gsub(/\A\\A/, '^').gsub(/\\z\z/, '$'), query.options)
 
-      files_to_scan = ls_tree(repo, nil, options.fetch(:ref, 'HEAD'), ls_tree_options)
+      ref = options.fetch(:ref, 'HEAD')
+      files_to_scan = ls_tree(repo, nil, ref, ls_tree_options)
 
       files_to_scan.each_with_object({}) do |file, result|
         id = if file[:mode] == SYMLINK_TYPE
@@ -231,7 +225,7 @@ module RJGit
             dir[-1] = symlink_source
             symlink_source = File.join(dir)
           end
-          Blob.find_blob(repo, symlink_source).jblob.id
+          Blob.find_blob(repo, symlink_source, ref).jblob.id
         else
           ObjectId.from_string(file[:id])
         end