I discovered that Porcelain.grep performs horribly.  I did some profiling and identified a few issues, mostly involving unnecessarily retrieving data that we already had in memory.

I tested performance against a repo of mine and a085d06bcc5ddd6880de479250f55586cfae0935 improved runtime of a simple `Porcelain.grep` call from 6.1 seconds to 0.57 seconds.

Commit d5019e6a67453a0191408067bcb9c53436c9ee25 further improved runtime to 0.28 seconds.  (The performance bottleneck that d5019e6a67453a0191408067bcb9c53436c9ee25 bypasses is `Java::OrgEclipseJgitTreewalk::TreeWalk.forPath`.)

The last remaining bottleneck was `get_bytes.to_a`, accounting for 80% of runtime.  b3d01eca9f3b802178ba77646ed5c46e05743623 fixes this.  I ran this against a much larger repo, and b3d01eca9f3b802178ba77646ed5c46e05743623 alone cut runtime from 103 seconds to 7 seconds!