@betesh 2020-06-03
@bartkamphorst Please take a look--this PR improves performance of `Porcelain.grep` by 99.7%