@bartkamphorst 2020-06-03
Impressive speed-up @betesh ! Unfortunately I won't have time to review this PR this week. I'll try to get to it after the weekend (or @dometto ?).