@betesh 2020-11-03
I ran `Porcelain.grep` on a large repo (43619 files) to get some new metrics.

On master: 4138.94 seconds
On this branch (both before and after addressing PR feedback): 10.55 seconds

That's a 99.75% improvement