@betesh 2020-11-05
@bartkamphorst  Sorry it took me so long to get to this.  I believe I've addressed all your feedback.