@dometto 2020-11-17
> On master: 4138.94 seconds (i.e. 1 hour, 9 minutes)
> On this branch (both before and after addressing PR feedback): 10.55 seconds
> That's a 99.75% improvement

Amazing result! Thanks so much for working on this.